Minor fixups/bugfixes. Only mangle imported tables.

+++ util.c
@@ -85,7 +85,7 @@ uint8_t *mmap_file(char *path, uint32_t *len) {
 void *acalloc(size_t nmemb, size_t size,  char *name) {
     void *res = calloc(nmemb, size);
     if (res == NULL) {
-        FATAL("Could not allocate %ld bytes for %s", nmemb * size, name);
+        FATAL("Could not allocate %d bytes for %s", (int)nmemb * size, name);
     }
     return res;
 }
@@ -95,7 +95,7 @@ void *arecalloc(void *ptr, size_t old_nmemb, size_t nmemb,
                 size_t size,  char *name) {
     void *res = realloc(ptr, nmemb * size);
     if (res == NULL) {
-        FATAL("Could not allocate %ld bytes for %s", nmemb * size, name);
+        FATAL("Could not allocate %d bytes for %s", (int)nmemb * size, name);
     }
     // Initialize new memory
     memset(res + old_nmemb * size, 0, (nmemb - old_nmemb) * size);
